Week 9/flask_app/requirements.txt
flask>=3.0.0
torch>=2.0.0
pillow>=10.0.0
